+++ MercadoPagoSDK/MercadoPagoSDK/Flows/InitFlow/MercadopagoCheckoutViewModel+InitFlow.swift
@@ -24,7 +24,7 @@ extension MercadoPagoCheckoutViewModel {
         initFlowProperties.paymentConfigurationService = self.paymentConfigurationService
         initFlowProperties.escManager = escManager
         initFlowProperties.privateKey = privateKey
-        initFlowProperties.productId = self.getAdvancedConfiguration().productId
+        initFlowProperties.productId = getAdvancedConfiguration().productId
 
         configureBiometricModule()
